requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
alsa-lib=1.2.8=h166bdaf_0
aom=3.5.0=h27087fc_0
attr=2.5.1=h166bdaf_1
blas=1.0=mkl
box2d-py=2.3.8=py39h5a03fae_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.5.7=hbcca054_0
cairo=1.16.0=h35add3b_1015
cffi=1.15.0=py39h4bc2ebd_0
cloudpickle=2.2.1=pyhd8ed1ab_0
dav1d=1.0.0=h166bdaf_1
dbus=1.13.6=h5008d03_3
expat=2.5.0=hcb278e6_1
filelock=3.12.0=pyhd8ed1ab_0
fluidsynth=2.3.2=hea804a5_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=hca18f0e_1
future=0.18.3=pyhd8ed1ab_0
gettext=0.21.1=h27087fc_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
graphite2=1.3.13=h58526e2_1001
gym=0.26.1=py39hcca971b_0
gym-box2d=0.26.1=py39hf3d152e_0
gym-notices=0.0.8=pyhd8ed1ab_0
harfbuzz=7.2.0=hdb3a94d_0
icu=72.1=hcb278e6_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46305
jack=1.9.22=h11f4161_0
jinja2=2.11.3=pyhd8ed1ab_2
jpeg=9e=h0b41bf4_3
lame=3.100=h166bdaf_1003
ld_impl_linux-64=2.38=h1181459_1
lerc=4.0.0=h27087fc_0
libavif=0.11.1=h5cdd6b5_0
libcap=2.67=he9d0100_0
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.17=h0b41bf4_0
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h6a678d5_6
libflac=1.4.2=h27087fc_0
libgcc-ng=12.2.0=h65d4601_19
libgcrypt=1.10.1=h166bdaf_0
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libglib=2.76.2=hebfc3b9_0
libgpg-error=1.46=h620e276_0
libiconv=1.17=h166bdaf_0
libmad=0.15.1b=h0b41bf4_1001
libogg=1.3.4=h7f98852_1
libopenblas=0.3.20=pthreads_h78a6416_0
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h753d276_0
libprotobuf=3.20.3=he621ea3_0
libsndfile=1.2.0=hb75c966_0
libstdcxx-ng=12.2.0=h46fd767_19
libsystemd0=253=h8c4010b_1
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.3.0=h0b41bf4_0
libxcb=1.13=h7f98852_1004
libzlib=1.2.13=h166bdaf_4
llvm-openmp=16.0.3=h4dfa4b3_0
lz4-c=1.9.4=hcb278e6_0
markupsafe=1.1.1=py39h3811e60_3
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpg123=1.31.3=hcb278e6_0
mpmath=1.3.0=pyhd8ed1ab_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pyhd8ed1ab_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.24.3=py39hf6e8229_1
numpy-base=1.24.3=py39h060ed82_1
openssl=1.1.1t=h7f8727e_0
opusfile=0.12=hae46d1a_2
pcre2=10.40=hc3806b6_0
pip=23.1.2=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
portaudio=19.6.0=h583fa2b_7
portmidi=2.0.4=h11f4161_0
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio-client=16.1=h5195f5e_3
pycparser=2.21=pyhd8ed1ab_0
pygame=2.1.3=py39ha038556_0
python=3.9.16=h7a1cb2a_2
python_abi=3.9=2_cp39
pytorch=1.13.1=cpu_py39h92724a6_0
pytorch-mutex=1.0=cpu
pyyaml=6.0=py39hb9d737c_4
readline=8.2=h5eee18b_0
sdl2=2.26.5=h949db6a_0
sdl2_image=2.6.3=h7b3d2e9_0
sdl2_mixer=2.6.3=h563c3bf_0
sdl2_ttf=2.20.2=h2123e47_1
setuptools=67.7.2=pyhd8ed1ab_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=pyh04b8f61_3
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
wheel=0.40.0=pyhd8ed1ab_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.8.4=h0b41bf4_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=h3eb15da_6
